Add files via upload
Capstone Functions is R functions to be used for ensemble forecast adjustment. Copstone Function Testing (Spelling error sorry) is the file that runs the functions. There are still some more things that will be added soon, such as an adjustment to the input data, adjustment to the output (switch to long format) etc. but these will be added later this week.